+++ students/ws/users/urls.py
@@ -3,5 +3,5 @@ from django.urls import path
 from users.views import SighUpView
 
 urlpatterns = [
-    path('/signup', SighUpView.as_view()),
+    path('/signup', SighUpView.as_view())
 ]
