+++ students/ws/users/views.py
@@ -1,19 +1,10 @@
-import re
 import json
 
-from django.http import JsonResponse
-from django.views import View
-from django.core.exceptions import ValidationError
+from django.http     import JsonResponse
+from django.views    import View
 
-from users.models import User
-
-def validate_email(email):
-    EMAIL_REGEX = re.match('^[a-zA-z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9.]+$', email)
-    return EMAIL_REGEX
-
-def validate_password(password):
-    PASSWORD_REGEX = re.match('^(?=.*[A-Za-z])(?=.*\d)(?=.*[$@$!%*#?&])[A-Za-z\d$@$!%*#?&]{8,}$', password)
-    return PASSWORD_REGEX
+from users.models    import User
+from users.validator import validate_email, validate_password
 
 class SighUpView(View):
     def post(self, request):
@@ -25,6 +16,9 @@ class SighUpView(View):
             email        = data['email']
             password     = data['password']
             phone_number = data['phone_number']
+
+            validate_email(email)
+            validate_password(password)
 
             if not validate_email(email):
                 return JsonResponse( {"message" : "INVALID_EMAIL"}, status = 400)
@@ -44,6 +38,6 @@ class SighUpView(View):
                 phone_number = phone_number
             )
 
-            return JsonResponse({"message": "SUCCESS"}, status=201)
+            return JsonResponse({"message": "SIGHUP SUCCESS"}, status=201)
         except KeyError:
             return JsonResponse( {"message" : "KEYERROR"}, status = 400)
